# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  global_name?: string | null;
  isAdmin: boolean;
  isAuthorized: boolean;
}

export interface AuthResult {
  user: DiscordUser;
  redirectUrl: string;
}

export interface AuthorizedUser {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  authorizedAt: string;
  authorizedBy: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tokenHash?: string | null;
  /** @nullable */
  lastSeen?: string | null;
  sessionCount: number;
}

export interface AuthorizeUserInput {
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface AccessCheckResult {
  authorized: boolean;
  isAdmin: boolean;
}

export type AutoAdSessionStatus = typeof AutoAdSessionStatus[keyof typeof AutoAdSessionStatus];


export const AutoAdSessionStatus = {
  idle: 'idle',
  running: 'running',
  stopped: 'stopped',
  error: 'error',
} as const;

export interface AutoAdSession {
  id: number;
  discordId: string;
  channelId: string;
  message: string;
  delay: number;
  interval: number;
  status: AutoAdSessionStatus;
  messagesSent: number;
  createdAt: string;
  /** @nullable */
  lastSentAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface SessionInput {
  channelId: string;
  message: string;
  delay: number;
  interval: number;
  userToken: string;
}

export interface TokenInput {
  token: string;
}

export interface TokenValidationResult {
  valid: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface ActivityItem {
  discordId: string;
  username: string;
  action: string;
  timestamp: string;
}

export interface AdminStats {
  totalAuthorized: number;
  activeSessionsCount: number;
  totalMessagesSent: number;
  recentActivity: ActivityItem[];
}

export interface MessageResponse {
  message: string;
}

export type DiscordCallbackParams = {
code: string;
state?: string;
};