# lib/api-zod/src/generated/types/accessCheckResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCheckResult {
  authorized: boolean;
  isAdmin: boolean;
}