# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  discordId: string;
  username: string;
  action: string;
  timestamp: string;
}