# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface AdminStats {
  totalAuthorized: number;
  activeSessionsCount: number;
  totalMessagesSent: number;
  recentActivity: ActivityItem[];
}