# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscordUser } from './discordUser';

export interface AuthResult {
  user: DiscordUser;
  redirectUrl: string;
}