# lib/api-zod/src/generated/types/authorizeUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorizeUserInput {
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  notes?: string | null;
}