# lib/api-zod/src/generated/types/authorizedUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorizedUser {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  authorizedAt: string;
  authorizedBy: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tokenHash?: string | null;
  /** @nullable */
  lastSeen?: string | null;
  sessionCount: number;
}