# lib/api-zod/src/generated/types/autoAdSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */
import type { AutoAdSessionStatus } from './autoAdSessionStatus';

export interface AutoAdSession {
  id: number;
  discordId: string;
  channelId: string;
  message: string;
  delay: number;
  interval: number;
  status: AutoAdSessionStatus;
  messagesSent: number;
  createdAt: string;
  /** @nullable */
  lastSentAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}