# lib/api-zod/src/generated/types/discordUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  global_name?: string | null;
  isAdmin: boolean;
  isAuthorized: boolean;
}