# lib/api-zod/src/generated/types/tokenValidationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * boxfight auto ad API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenValidationResult {
  valid: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  error?: string | null;
}